requirements.txt
# Install PyTorch with CUDA 12.6 support:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu126
torch>=2.1.0
torchvision>=0.16.0

# Data
kagglehub>=1.0.0
kaggle>=2.1.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0

# Training utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.14.0
matplotlib>=3.7.0
